Finished adding TokenContext, currently it manages the token, but I intend to add more to it. Moved Token Selection components out of FundBounty into their own directory. Edited import paths and tests to accomodate this.
