--- components/FundBounty/TokenSelection/TokenFundBox/TokenFundBox.test.js
@@ -1,14 +1,0 @@
-/**
- * @jest-environment jsdom
- */
-import React from 'react';
-import TokenFundBox from '../../../components/FundBounty/TokenSelection/TokenFundBox';
-import Constants from '../../../test-utils/constant';
-import renderer from 'react-test-renderer';
-
-describe('TokenFundBox', () => {
-  it('should match DOM Snapshot', () => {
-    const tree = renderer.create(<TokenFundBox token={Constants.tokenMetadata} />);
-    expect(tree.toJSON()).toMatchSnapshot('shows token symbol, img and input field');
-  });
-});
