+++ components/MintBounty/MintBountyModal/AddSplitPriceParams.js
@@ -1,9 +1,9 @@
 import React, { useContext, useEffect } from 'react';
-import TokenFundBox from '../../FundBounty/TokenSelection/TokenFundBox';
+import TokenFundBox from '../../TokenSelection/TokenFundBox';
 import { ethers } from 'ethers';
 import ToolTipNew from '../../Utils/ToolTipNew';
 import MintContext from '../MintContext';
-import TokenContext from '../../FundBounty/TokenSelection/TokenStore/TokenContext';
+import TokenContext from '../../TokenSelection/TokenStore/TokenContext';
 import { parseVolume } from '../../../services/utils/lib';
 
 const AddSplitPriceParams = () => {
