+++ components/TokenSelection/ManageTokenList/index.js
@@ -3,10 +3,10 @@ import React, { useContext, useState } from 'react';
 import Image from 'next/image';
 
 // Custom Hooks
-import Toggle from '../../../Utils/Toggle';
+import Toggle from '../../Utils/Toggle';
 import { ethers } from 'ethers';
-import StoreContext from '../../../../store/Store/StoreContext';
-import TokenDisplay from '../../../TokenBalances/TokenDisplay';
+import StoreContext from '../../../store/Store/StoreContext';
+import TokenDisplay from '../../TokenBalances/TokenDisplay';
 
 const ManageTokenList = ({ setCustomTokens, customTokens, setLists, lists, stream }) => {
   const [displayedTab, setDisplayedTab] = useState('Lists');
